refactor: change tests structure
- Changed directory structure
- Removed @online decorator to ensure that every tests is being run in
  the CI instead of being silently skipped
