+++ backend/kernelCI_app/tests/integrationTests/buildDetails_test.py
@@ -1,14 +1,13 @@
-from kernelCI_app.unitTests.utils.healthCheck import online
-from kernelCI_app.unitTests.utils.client.buildClient import BuildClient
-from kernelCI_app.unitTests.utils.asserts import (
+from kernelCI_app.tests.utils.client.buildClient import BuildClient
+from kernelCI_app.tests.utils.asserts import (
     assert_status_code_and_error_response,
     assert_has_fields_in_response_content,
 )
-from kernelCI_app.unitTests.utils.fields.builds import (
+from kernelCI_app.tests.utils.fields.builds import (
     build_details_expected_fields,
     build_tests_expected_fields,
 )
-from kernelCI_app.unitTests.utils.fields.issues import issues_resource_fields
+from kernelCI_app.tests.utils.fields.issues import issues_resource_fields
 from kernelCI_app.utils import string_to_json
 import pytest
 from http import HTTPStatus
@@ -17,7 +16,6 @@ from http import HTTPStatus
 client = BuildClient()
 
 
-@online
 @pytest.mark.parametrize(
     "build_id, status_code, has_error_body",
     [
@@ -43,7 +41,6 @@ def test_get_build_details(
         )
 
 
-@online
 @pytest.mark.parametrize(
     "build_id, status_code, has_error_body",
     [
@@ -76,7 +73,6 @@ def test_get_build_tests(
                 )
 
 
-@online
 @pytest.mark.parametrize(
     "build_id, status_code, has_error_body",
     [
