+++ backend/kernelCI_app/tests/integrationTests/hardwareDetailsSummary_test.py
@@ -3,13 +3,12 @@ from requests import Response
 from http import HTTPStatus
 
 from kernelCI_app.utils import string_to_json
-from kernelCI_app.unitTests.utils.healthCheck import online
-from kernelCI_app.unitTests.utils.client.hardwareClient import HardwareClient
-from kernelCI_app.unitTests.utils.commonTreeAsserts import (
+from kernelCI_app.tests.utils.client.hardwareClient import HardwareClient
+from kernelCI_app.tests.utils.commonTreeAsserts import (
     assert_common_summary_status_fields,
 )
-from kernelCI_app.unitTests.utils.fields import hardware
-from kernelCI_app.unitTests.utils.asserts import (
+from kernelCI_app.tests.utils.fields import hardware
+from kernelCI_app.tests.utils.asserts import (
     assert_has_fields_in_response_content,
     assert_status_code,
     assert_error_response,
@@ -160,7 +159,6 @@ def request_data(
     return response, content
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, status_code, has_error_body",
     [
@@ -203,7 +201,6 @@ def test_no_filters(base_hardware, status_code, has_error_body):
         )
 
 
-@online
 def test_filter_test_status(test_status_input):
     """
     Tests for the status filter for both boots and tests
@@ -235,7 +232,6 @@ def test_filter_test_status(test_status_input):
                 assert count == 0
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, filters",
     [
@@ -267,7 +263,6 @@ def test_filter_build_status(base_hardware, filters):
                 assert count == 0
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, filters",
     [
@@ -291,7 +286,6 @@ def test_filter_config_name(base_hardware, filters):
             assert config in values
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, filters",
     [
@@ -322,7 +316,6 @@ def test_filter_architectures(base_hardware, filters):
         assert arch in values
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, filters",
     [
@@ -354,7 +347,6 @@ def test_filter_compiler(base_hardware, filters):
             assert compiler in values
 
 
-@online
 @pytest.mark.parametrize(
     "base_hardware, filters",
     [
@@ -388,7 +380,6 @@ def test_platform(base_hardware, filters):
         assert task_summary["status"][status] == count
 
 
-@online
 def test_filter_issues(issues_input):
     base_hardware, filters = issues_input
     response, content = request_data(base_hardware, filters)
@@ -419,7 +410,6 @@ def test_filter_issues(issues_input):
     assert pass_count == 0
 
 
-@online
 def test_invalid_filters(invalid_filters_input):
     empty_test = {
         "architectures": [],
