+++ backend/kernelCI_app/tests/integrationTests/testDetails_test.py
@@ -1,11 +1,10 @@
-from kernelCI_app.unitTests.utils.healthCheck import online
-from kernelCI_app.unitTests.utils.client.testClient import TestClient
-from kernelCI_app.unitTests.utils.asserts import (
+from kernelCI_app.tests.utils.client.testClient import TestClient
+from kernelCI_app.tests.utils.asserts import (
     assert_status_code_and_error_response,
     assert_has_fields_in_response_content,
 )
-from kernelCI_app.unitTests.utils.fields.tests import test_expected_fields
-from kernelCI_app.unitTests.utils.fields.issues import issues_resource_fields
+from kernelCI_app.tests.utils.fields.tests import test_expected_fields
+from kernelCI_app.tests.utils.fields.issues import issues_resource_fields
 from kernelCI_app.utils import string_to_json
 import pytest
 from http import HTTPStatus
@@ -14,7 +13,6 @@ from http import HTTPStatus
 client = TestClient()
 
 
-@online
 @pytest.mark.parametrize(
     "test_id, status_code, has_error_body",
     [
@@ -38,7 +36,6 @@ def test_get(test_id, status_code, has_error_body):
         )
 
 
-@online
 @pytest.mark.parametrize(
     "test_id, status_code, has_error_body",
     [
