+++ backend/kernelCI_app/tests/utils/client/buildClient.py
@@ -1,7 +1,7 @@
 import requests
 from django.urls import reverse
 
-from kernelCI_app.unitTests.utils.client.baseClient import BaseClient
+from kernelCI_app.tests.utils.client.baseClient import BaseClient
 
 
 class BuildClient(BaseClient):
