+++ backend/kernelCI_app/tests/utils/fields/hardware.py
