+++ backend/kernelCI_app/tests/utils/hardwareDetailsCommonTestCases.py
